requirements.txt
pillow>=10.0.0
pydub>=0.25.1

